Specify stylelint as a peer dependencies
This follows the common approach used in stylelint rules of not setting
the project as a direct dependency and suggesting the peer installs it.

Looking on [awesome-stylelint](https://github.com/stylelint/awesome-stylelint)
the convention for configurations is to only set stylelint as a peer
dependency and not set any plugins as this, this is why I've not the
same for stylelint-scss.