Disable no-descending-specificity
This rule wasn't easy to apply to our code base because we have
components that use BEM modified nestings and the selector may be at a
different specificity.

For example:

```

  .govuk-checkboxes__input:disabled,
  .govuk-checkboxes__input:disabled + .govuk-checkboxes__label {
    cursor: default;
  }

  .govuk-checkboxes--small {
    .govuk-checkboxes__label {
      margin-top: -2px;
      padding: 13px govuk-spacing(3) 13px 1px;
      float: left;

      @include govuk-media-query($from: tablet) {
        padding: 11px govuk-spacing(3) 10px 1px;
      }
    }
  }
```

In the above example the later code is at a higher specificity, however
it would be unconventional for GDS component style to define the earlier
rule later in the file.