Require single colons for pseudo-element rules
This is changed is prompted from a discussion in
govuk_publishing_components [1] where the consensus was that we should
maintain the single colon approach to pseudo-elements for IE8
compatibility. The logic for this was that, while we may not support
IE8, syntax wasn't a sufficient reason to risk visual problems.

Presumably this can be removed once we reach a point of deeming IE8
insignificant enough to no longer worry about or if auto-prefixer is
ubiquitous (which can convert from double to single automatically in
compilation in IE8).

[1]: https://github.com/alphagov/govuk_publishing_components/pull/1611#discussion_r456497712