Add release-it for releasing this module
This adds the release-it module in to help provide a consistent release
process for this module. release-it is a module that manages various
aspects of a release process, such as setting version number, creating
changelog, tagging git, releasing to GitHub and publishing to NPM. This
is done since, for the time being, releasing needs to be a manual
process as there are unresolved issues that prevent us from using CI
(GitHub Actions) to release the module as is normally done with GOV.UK
modules.

I've had to somewhat bend the will of release-it to match the GOV.UK
working patterns. We tend to release a master branch after merging in
the version number and have branch protection that stops pushing to the
default branch - which ruled out some of release-it's functionality such
as setting a version number (disabled by increment: false) and pushing
the new release to the default branch. We also tend to maintain our own
curated changelogs.

Given this, the usage of release-it is intended to occur after a merge
has been made to set the new version number. Once this is done
release-it can be run and this will: tag the version in git, create a
release in GitHub (with a changelog based off commits after the last
tag), and publish the release to NPM. It also does some basic checks
that you're on the appropriate branch (main) and have a clean working
directory.

One thing to note is that I had to replace the default command to
generate a changelog from git history for GitHub releases. The way
release-it expects to operate is that it gives you a new version number
and tags the repo, then determines a changelog based on 2nd to last to
tag to newest tag. This didn't work for us when we're doing the tagging
after setting the version so I adapted the command to pull out all git
commits from last tag until current commit.