Resolve multiple inheritance challenges
This resolves a problem that the multiple inheritance of scss.js was
overwriting aspects of stylelint-config-standard-scss. I resolved this
by breaking the files up so there wasn't multiple degrees of inheritance
considered.

I am a bit concerned by the complexity of considering things like
stylelint-config-standard-scss, where there are 6 files rules could be
coming from for SCSS:

- stylelint-config-recommended
- stylelint-config-standard
- stylelint-config-recommended-scss
- stylelint-config-standard-scss
- css-rules
- scss-rules

That seems pretty brutal to understand, but does build on a lot of
community rules :-/. I'm going to try this out and may reign it back in
if it becomes a source of problem.