Disable declaration-block-no-redundant-longhand-properties
This rules has been disabled because of concerns raised by developers in
the GOV.UK Design System in: https://github.com/alphagov/govuk-frontend/pull/2567
where it may produce less readable code.

There was some confusion as to quite why it was added to
stylelint-config-standard [1] as the concerns for why it wasn't
previously a part of stylelint-config-standard [2] still seem to apply.

[1]: https://github.com/stylelint/stylelint-config-standard/blob/main/CHANGELOG.md#2300
[2]: https://github.com/stylelint/stylelint-config-standard/issues/113