Remove `selector-pseudo-element-colon-notation`
The Design System team is in the process of removing support for Internet Explorer 8, 9 and 10 from GOV.UK Frontend, signalling our intent to no longer support the browser or provide fixes or polyfills specifically for it.

Supporting IE8–10 has not been a requirement for services since June 2018.

Barring legacy projects, we do not expect services to be making accommodations for legacy IE versions now or in future, and encourage those projects to begin removing any code specifically for these versions of IE.

As part of this, we should remove `selector-pseudo-element-colon-notation` to begin enforcing the default standards-compliant double-colon syntax instead.